osbuild.spec: use forge macros
Defining `commit` when building the rpm will now expect the sources to
be in `osbuild-%{commit}.tar.gz` and also include the commit has in the
Release field. This makes building an RPM from an unreleased version
less confusing and allows multiple such builds to live in the same
repository.

Everything should be as before When not specifying `commit`.